refactor(extractor): moving older (unusable) tests
